add better exception testing

+++ src/Managers/IProductManager.cs
@@ -4,5 +4,5 @@ using System.Threading.Tasks;
 public interface IProductManager
 {
     public Task<Product> GetProductAsync(long id);
-    public Task<bool> UpdateProductPriceAsync(long id, decimal price);
+    public Task<bool> UpdateProductPriceAsync(long id, double price);
 }
