+++ src/Repositories/IProductRepository.cs
@@ -4,5 +4,5 @@ public interface IProductRepository
 {
     public Task<Product> GetProductAsync(long id);
     public Task<Price> GetProductPriceAsync(long id);
-    public Task<bool> UpdateProductPriceAsync(long id, decimal price);
+    public Task<bool> UpdateProductPriceAsync(long id, double price);
 }
